Removed functions from load sizing and moved them over to methods for level 2. This solves a circular import issue
